Move deploy method to manager
